Moved Example within main workspace
