--- Example/SDKExample/Podfile
@@ -1,9 +1,0 @@
-platform :ios, '9.0'
-
-use_frameworks!
-inhibit_all_warnings!
-
-target 'SDKExample' do
-  pod 'Drift', :path => '../../Drift.podspec'
-
-end
