--- Example/SDKExample/SDKExample/AppDelegate.swift
@@ -1,27 +1,0 @@
-//
-//  AppDelegate.swift
-//  SDKExample
-//
-//  Created by Eoin O'Connell on 29/05/2017.
-//  Copyright © 2017 drift. All rights reserved.
-//
-
-import UIKit
-import Drift
-
-@UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
-
-    var window: UIWindow?
-    
-    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
-
-        Drift.setup(<#embedId#>)
-        
-//        Drift.debugMode(true)
-        Drift.registerUser(<#userId#>, email: <#email#>)
-        
-        return true
-    }
-}
-
