+++ Podfile
@@ -4,10 +4,13 @@ use_frameworks!
 target 'Drift-SDK' do
     pod 'ObjectMapper', '~> 3.0'
     pod 'AlamofireImage', '~> 3.3'
-    pod 'SVProgressHUD', '~> 2.0'
     pod 'Starscream', '3.1.1'
 end
 
 target 'DriftTests' do
     pod 'ObjectMapper', '~> 3.0'
+end
+
+target 'SDKExample' do
+    pod 'Drift', :path => '.'
 end
