+++ SDKExample/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -31,15 +31,13 @@
       "scale" : "3x"
     },
     {
-      "size" : "60x60",
       "idiom" : "iphone",
-      "filename" : "app-icon 60x60@2x.png",
+      "size" : "60x60",
       "scale" : "2x"
     },
     {
-      "size" : "60x60",
       "idiom" : "iphone",
-      "filename" : "app-icon 60x60@3x.png",
+      "size" : "60x60",
       "scale" : "3x"
     },
     {
